+++ Project/GameEngine/Source/WADMaker.cpp
@@ -52,7 +52,7 @@ void WADMaker::Make(strgv inpath, strgv outloc, bool saveExt)
 
 	std::ofstream OUT(soutloc + outname, std::ios::binary);
 
-	uint fileCnt = filenames.size();
+	uint fileCnt = static_cast<uint>(filenames.size());
 	OUT.write(reinterpret_cast<const char*>(&fileCnt), sizeof(fileCnt));
 
 	uint currentOffset = sizeof(fileCnt);
@@ -60,7 +60,7 @@ void WADMaker::Make(strgv inpath, strgv outloc, bool saveExt)
 	for (uint i = 0; i < fileCnt; ++i)
 	{
 		currentOffset += sizeof(uint16);
-		currentOffset += filenames[i].size();
+		currentOffset += static_cast<uint>(filenames[i].size());
 		currentOffset += sizeof(uint);
 		currentOffset += sizeof(uint);
 		currentOffset += sizeof(uint8);
@@ -68,8 +68,8 @@ void WADMaker::Make(strgv inpath, strgv outloc, bool saveExt)
 
 	for (uint i = 0; i < fileCnt; ++i)
 	{
-		uint16 nameLength = filenames[i].size();
-		uint fileSize = data[i].size();
+		uint16 nameLength = static_cast<uint>(filenames[i].size());
+		uint fileSize = static_cast<uint>(data[i].size());
 		uint fileOffset = currentOffset;
 		uint8 fileType = filetypes[i];
 
@@ -106,7 +106,7 @@ void WADMaker::MakeUnique(strgv filename, strgv outloc, std::vector<WADMemItem>&
 	strg OUTPATH = soutloc + strg(filename);
 	std::ofstream OUT(OUTPATH.data(), std::ios::binary);
 
-	uint fileCnt = data.size();
+	uint fileCnt = static_cast<uint>(data.size());
 	OUT.write(reinterpret_cast<const char*>(&fileCnt), sizeof(fileCnt));
 
 	uint currentOffset = sizeof(fileCnt);
@@ -114,7 +114,7 @@ void WADMaker::MakeUnique(strgv filename, strgv outloc, std::vector<WADMemItem>&
 	for (auto& it : data)
 	{
 		currentOffset += sizeof(uint16);
-		currentOffset += it.Name.size();
+		currentOffset += static_cast<uint>(it.Name.size());
 		currentOffset += sizeof(uint);
 		currentOffset += sizeof(uint);
 		currentOffset += sizeof(uint8);
@@ -122,8 +122,8 @@ void WADMaker::MakeUnique(strgv filename, strgv outloc, std::vector<WADMemItem>&
 
 	for (auto& it : data)
 	{
-		uint16 nameLength = it.Name.size();
-		uint fileSize = it.Data.size();
+		uint16 nameLength = static_cast<uint16>(it.Name.size());
+		uint fileSize = static_cast<uint>(it.Data.size());
 		uint fileOffset = currentOffset;
 		uint8 fileType = (uint8)it.Item.Type;
